requirements.txt
email-validator
fastapi
jose
uvicorn
pydantic
python-dotenv
passlib
psycopg2
PyJWT
requests
sqlalchemy